slack_bot/requirements.txt
slack-bolt
fastapi
uvicorn
python-dotenv
aiohttp
mem0ai
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0
